% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenbrown.R
\docType{data}
\name{greenbrown}
\alias{greenbrown}
\title{Green vs Brown portfolio dataset}
\format{
A data frame with 2266 rows and columns:
\describe{
\item{DATE}{Dates going from 2014-01-02 to 2024-12-30.}
\item{return_green}{Numeric returns for the green portfolio.}
\item{return_brown}{Numeric returns for the brown portfolio.}
}
}
\source{
Included originally as \code{inst/extdata/green-brown-ptf.xlsx}.
}
\usage{
data(greenbrown)
}
\description{
Daily returns for a green and a brown portfolio constructed following the equal 1090
methodology.
}
\examples{
data("greenbrown")
str(greenbrown)
head(greenbrown)

}
\keyword{datasets}
