% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenbrown.R
\docType{data}
\name{greenbrown}
\alias{greenbrown}
\title{Green vs Brown portfolio dataset}
\format{
A data frame with 2266 rows and three columns:
\describe{
\item{DATE}{Dates ranging from 2014-01-02 to 2024-12-30.}
\item{return_green}{Numeric returns for the green portfolio.}
\item{return_brown}{Numeric returns for the brown portfolio.}
}
}
\source{
Originally data in \code{inst/extdata/green-brown-ptf.xlsx}.
}
\usage{
data(greenbrown)
}
\description{
Daily returns for a green and a brown portfolios constructed following
the equal-weighted 10-90 percentile approach.
}
\examples{
data("greenbrown")
str(greenbrown)
head(greenbrown)

}
\keyword{datasets}
