% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdc.R
\name{RSDC}
\alias{RSDC}
\alias{RSDC-package}
\title{RSDC: Regime-Switching Correlation Models for Portfolio Analysis}
\description{
\code{RSDC} is an R package for modeling and comparing correlation structures across regimes in multivariate time series.
It implements a regime-switching correlation framework with or without exogenous variables in the transition probabilities
(Time-Varying Transition Probabilities – TVTP), and allows full pipeline integration including model fitting, covariance
reconstruction, and portfolio optimization (e.g., minimum variance, maximum diversification).

The package is designed for empirical asset pricing and portfolio applications, and supports the inclusion of predictors
such as macroeconomic or climate-related indices to investigate time-varying market regimes.
}
\note{
By using \code{RSDC}, you agree to the following:
(1) You must include a link to \url{https://CRAN.R-project.org/package=RSDC} in any publication or documentation;
(2) You assume full responsibility for the interpretation and use of the results.
}
\section{Functions}{

\itemize{
\item Model estimation and filtering: \code{\link{rsdc_hamilton}}, \code{\link{rsdc_likelihood}}, \code{\link{rsdc_estimate}};
\item Correlation forecasting: \code{\link{rsdc_forecast}};
\item Portfolio construction: \code{\link{rsdc_minvar}}, \code{\link{rsdc_maxdiv}};
\item Complete simulation pipeline: \code{\link{rsdc_simulate}};
}
}

\section{Update}{

The latest version of the package is available at \url{https://github.com/ArdiaD/RegimeSwitchingCorrelation}
}

\references{
Hamilton, J. D. (1989). A new approach to the economic analysis of nonstationary time series and the business cycle.
\emph{Econometrica}, \bold{57}(2), 357--384.
}
\author{
David Ardia and Benjamin Seguin
}
